requirements.txt
# Pipeline + API
python-dotenv>=1.0.0
groq>=0.4.0
openai>=1.0.0
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pydantic>=2.0.0
httpx>=0.27.0

# Data sources (Kaggle, Apify, scraping)
kaggle>=1.6.0
apify-client>=1.6.0
beautifulsoup4>=4.12.0
pandas>=2.0.0

# Solara app (optional)
solara>=1.0.0
plotly>=5.0.0
